fix: allow complex objects to be used in the notSetValue shorthand
The previous test for notSetValue shorthand was just typeof object.
This meant that arrays and null and dates and objects would fail the
test and the record would then try to find .notSetValue on that object.
This would most often be undefined, but would have been an error for
null.

This change makes the test check for isKeyed and then has key of
'notSetValue'. This is still not amazing as getters and setters will
have to explicitly define a notSetValue of undefined if that's their use
case. But we can't really check for `get` or `set` as it is likely that
default values would contain those keys.